:recycle: 構成整理
