+++ apps/frontend/src/features/news/components/NewsList.tsx
@@ -1,0 +1,3 @@
+export const NewsList = (): JSX.Element => {
+  return <></>;
+};
