+++ apps/frontend/src/features/news/components/index.ts
@@ -1,3 +1,2 @@
-export * from "./News";
-export * from "./NewsCard";
-export * from "./NewsCardList";
+export * from "./NewsPreview";
+export * from "./NewsList";
