+++ apps/frontend/src/features/news/types/index.ts
@@ -1,4 +1,4 @@
-export interface News {
+export interface NewsType {
   thumbnailSrc: string;
   title: string;
   updateDate: string;
