+++ apps/frontend/src/pages/index.tsx
@@ -2,7 +2,7 @@ import { MainLayout } from "@/components/Layout";
 import { AboutTeam } from "@/features/about-team/components";
 import { HeroContent } from "@/features/hero-content/components";
 import { JuniorYouth } from "@/features/junior-youth/components";
-import { News } from "@/features/news/components";
+import { NewsPreview } from "@/features/news/components";
 import { SocialContribution } from "@/features/social-contribution/components";
 import { SponsorRecruiting } from "@/features/sponsor-recruiting/components";
 import { NextPage } from "next";
@@ -28,7 +28,7 @@ const Top: NextPage = () => {
         </section>
 
         <section className="mt-[220px] sm:mt-[320px]">
-          <News />
+          <NewsPreview />
         </section>
 
         <section className="mt-[320px]">
