+++ apps/frontend/src/pages/news.tsx
@@ -1,10 +1,11 @@
 import { MainLayout } from "@/components/Layout";
+import { NewsList } from "@/features/news/components";
 import { NextPage } from "next";
 
 const News: NextPage = () => {
   return (
     <MainLayout>
-      <div>News</div>
+      <NewsList />
     </MainLayout>
   );
 };
